refactor(commands): extract bug/history tests to sibling file
